requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
aiohttp>=3.9.0
asyncio>=3.4.3
requests>=2.31.0
Pillow>=10.0.0
psutil>=5.9.0
flask>=2.3.0
flask-cors>=4.0.0

